requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
crispy-bootstrap5==0.5
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.17.0
django-colorfield==0.4.3
django-crispy-forms==1.12.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.2
pipreqs==0.4.10
psycopg2==2.9.1
python-interface==1.6.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
yarg==0.1.9
